Documentation for how to run the site
Added documentation for development and deployment of the site
Added instructions for adding new events, sponsors and teammates

Made event calendar and sponsor sections dependent on there existing
any events. It will be redundant after we start events but a nice way
to make sure that adding something to a data file is all we need to keep
the website up to date once we confirm the first event.